requirements.txt
ansible==2.8.6
docker-py==1.10.6
molecule==2.22
python-jenkins==1.5.0
testinfra==3.2.0
